+++ routes/index.ts
@@ -1,7 +1,12 @@
 import './signMessage'
-import './genKeysAndSession'
-import './SignWithSession'
+import './middleware'
+import './makeBlake'
+import './generateCid'
+import './genKeys'
+import './signMessage'
+import './signArbitrary'
 import './provisionSession'
+
 import { app } from '../index'
 
 export { app }
